Beginning refactoring and code quality improvements
Starting to rewrite all the JS in ES6. Previously the code was quite
messy so I’m in the process of improving it.

Moved the initialisation code into its own separate file
(scripts/app.js). This will be refactored in the hours to come.

Let’s hope he tests it on a modern browser (i.e. not IE) or else we’re
doomed